Add files via upload
Hello, there is an editor here that you can use, now I will tell you exactly how, there is a panel for choosing a color, the thickness of the cursor line, you can also import an image to modify it, be careful that the image must be selected up to 900 pixels in size, as they cost restrictions! In this editor, you can draw with LMB, also erase with RMB, be careful if you erase, then by default the 1st color that you will use will be black, to change it you will need to double-click on the color selection! ou can also save what you managed to draw somewhere to your local storage, or you can clear the canvas, you can try experimenting with this editor, all the best!